+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## master (unreleased)
 
+- Make background data migrations scheduler run a single migration at a time
+
+    This is similar to background schema migrations scheduler's behavior.
+    Running multiple migrations at a time can cause unneeded database load and other bad effects.
+
 - Add `#can_be_cancelled?` and `#can_be_paused?` helpers to background migrations
 
 ## 0.21.0 (2024-12-09)